requirements.txt
googletrans==3.1.0a0
matplotlib==3.10.7
numpy==2.3.4
opencv_contrib_python==4.10.0.84
opencv_python==4.12.0.88
paddleocr==3.2.0
Pillow==12.0.0
pypdfium2==4.30.0
ultralytics==8.3.213
